+++ GPT_SoVITS/qconfig.py
@@ -1,395 +1,13 @@
-# 来源：https://github.com/zhiyiYo/PyQt-Fluent-Widgets/blob/0e83ba1ef2e518dd4d66f0924a24be4d29a2cdbc/qfluentwidgets/common/config.py
-# 原仓库遵守 GPL-3.0 许可证，版权归 zhiyiYo 所有。
-# 因为本项目同样采用 GPL-3.0 许可证，可以直接使用该代码。
+# 数字小祥项目的统一配置类
+
 import json
 import os
-from copy import deepcopy
-from enum import Enum
-from pathlib import Path
-from typing import List
 import warnings
+import contextlib
 
-from PyQt5.QtCore import QObject, pyqtSignal
-from PyQt5.QtGui import QColor
 
-
-class ConfigValidator:
-    """ Config validator """
-
-    def validate(self, value) -> bool:
-        """ Verify whether the value is legal """
-        return True
-
-    def correct(self, value):
-        """ correct illegal value """
-        return value
-
-
-class RangeValidator(ConfigValidator):
-    """ Range validator """
-
-    def __init__(self, min, max):
-        self.min = min
-        self.max = max
-        self.range = (min, max)
-
-    def validate(self, value):
-        return self.min <= value <= self.max
-
-    def correct(self, value):
-        return min(max(self.min, value), self.max)
-
-
-class OptionsValidator(ConfigValidator):
-    """ Options validator """
-
-    def __init__(self, options):
-        if not options:
-            raise ValueError("The `options` can't be empty.")
-
-        if isinstance(options, Enum):
-            options = options._member_map_.values()
-
-        self.options = list(options)
-
-    def validate(self, value) -> bool:
-        """
-        Verify whether the value is legal by checking if it is in the options list
-        """
-        return value in self.options
-
-    def correct(self, value):
-        return value if self.validate(value) else self.options[0]
-
-
-class BoolValidator(OptionsValidator):
-    """ Boolean validator, only allow True and False option values """
-
-    def __init__(self):
-        super().__init__([True, False])
-
-
-class FolderValidator(ConfigValidator):
-    """ Folder validator """
-
-    def validate(self, value):
-        return Path(value).exists()
-
-    def correct(self, value):
-        path = Path(value)
-        path.mkdir(exist_ok=True, parents=True)
-        return str(path.absolute()).replace("\\", "/")
-
-
-class FolderListValidator(ConfigValidator):
-    """ Folder list validator """
-
-    def validate(self, value):
-        return all(Path(i).exists() for i in value)
-
-    def correct(self, value: List[str]):
-        folders = []
-        for folder in value:
-            path = Path(folder)
-            if path.exists():
-                folders.append(str(path.absolute()).replace("\\", "/"))
-
-        return folders
-
-
-class ColorValidator(ConfigValidator):
-    """ RGB color validator """
-
-    def __init__(self, default):
-        self.default = QColor(default)
-
-    def validate(self, value) -> bool:
-        try:
-            return QColor(value).isValid()
-        except:
-            return False
-
-    def correct(self, value):
-        return QColor(value) if self.validate(value) else self.default
-
-
-class ConfigSerializer:
-    """ Config serializer """
-
-    def serialize(self, value):
-        """ serialize config value """
-        return value
-
-    def deserialize(self, value):
-        """ deserialize config from config file's value """
-        return value
-
-
-class EnumSerializer(ConfigSerializer):
-    """ enumeration class serializer """
-
-    def __init__(self, enumClass):
-        self.enumClass = enumClass
-
-    def serialize(self, value):
-        return value.value
-
-    def deserialize(self, value):
-        return self.enumClass(value)
-
-
-class ColorSerializer(ConfigSerializer):
-    """ QColor serializer """
-
-    def serialize(self, value: QColor):
-        return value.name(QColor.HexArgb)
-
-    def deserialize(self, value):
-        if isinstance(value, list):
-            return QColor(*value)
-
-        return QColor(value)
-
-
-class ConfigItem(QObject):
-    """ Config item """
-
-    valueChanged = pyqtSignal(object)
-
-    def __init__(self, group: str, name: str, default, validator: ConfigValidator | None = None, serializer: ConfigSerializer | None = None, restart: bool = False):
-        """
-        Parameters
-        ----------
-        group: str
-            config group name
-
-        name: str
-            config item name, can be empty
-
-        default:
-            default value
-
-        options: list
-            options value
-
-        serializer: ConfigSerializer
-            config serializer
-
-        restart: bool
-            whether to restart the application after updating value
-        """
-        super().__init__()
-        self.group = group
-        self.name = name
-        self.validator = validator or ConfigValidator()
-        self.serializer = serializer or ConfigSerializer()
-        self.__value = default
-        self.value = default
-        self.restart = restart
-        self.defaultValue = self.validator.correct(default)
-
-    @property
-    def value(self):
-        """ get the value of config item """
-        return self.__value
-
-    @value.setter
-    def value(self, v):
-        v = self.validator.correct(v)
-        ov = self.__value
-        self.__value = v
-        if ov != v:
-            self.valueChanged.emit(v)
-
-    @property
-    def key(self):
-        """ get the config key separated by `.` """
-        return self.group+"."+self.name if self.name else self.group
-
-    def __str__(self):
-        return f'{self.__class__.__name__}[value={self.value}]'
-
-    def serialize(self):
-        return self.serializer.serialize(self.value)
-
-    def deserializeFrom(self, value):
-        self.value = self.serializer.deserialize(value)
-
-
-class RangeConfigItem(ConfigItem):
-    """ Config item of range """
-
-    @property
-    def range(self):
-        """ get the available range of config """
-        return self.validator.range
-
-    def __str__(self):
-        return f'{self.__class__.__name__}[range={self.range}, value={self.value}]'
-
-
-class OptionsConfigItem(ConfigItem):
-    """ Config item with options """
-
-    @property
-    def options(self):
-        return self.validator.options
-
-    def __str__(self):
-        return f'{self.__class__.__name__}[options={self.options}, value={self.value}]'
-
-
-class ColorConfigItem(ConfigItem):
-    """ Color config item """
-
-    def __init__(self, group, name, default, restart=False):
-        super().__init__(group, name, QColor(default), ColorValidator(default),
-                         ColorSerializer(), restart)
-
-    def __str__(self):
-        return f'{self.__class__.__name__}[value={self.value.name()}]'
-    
-
-def exceptionHandler(*default):
-    """ decorator for exception handling
-
-    Parameters
-    ----------
-    *default:
-        the default value returned when an exception occurs
-    """
-
-    def outer(func):
-
-        def inner(*args, **kwargs):
-            try:
-                return func(*args, **kwargs)
-            except BaseException as e:
-                value = deepcopy(default)
-                if len(value) == 0:
-                    return None
-                elif len(value) == 1:
-                    return value[0]
-
-                return value
-
-        return inner
-
-    return outer
-
-
-class QConfig(QObject):
-    """ Config of app """
-    # 当修改了一个被标记为 restart=True（需要重启后生效）的配置项时，发出该信号
-    appRestartSignal = pyqtSignal()
-
-    def __init__(self):
-        super().__init__()
-        self.file = Path("../d_sakiko_config.json")
-        self._cfg = self
-
-    def get(self, item):
-        """ get the value of config item """
-        return item.value
-
-    def set(self, item, value, save=True, copy=True):
-        """ set the value of config item
-
-        Parameters
-        ----------
-        item: ConfigItem
-            config item
-
-        value:
-            the new value of config item
-
-        save: bool
-            whether to save the change to config file
-
-        copy: bool
-            whether to deep copy the new value
-        """
-        if item.value == value:
-            return
-
-        # deepcopy new value
-        try:
-            item.value = deepcopy(value) if copy else value
-        except:
-            item.value = value
-
-        if save:
-            self.save()
-        
-        if item.restart:
-            self.appRestartSignal.emit()
-
-    def toDict(self, serialize=True):
-        """ convert config items to `dict` """
-        items = {}
-        for name in dir(self._cfg.__class__):
-            item = getattr(self._cfg.__class__, name)
-            if not isinstance(item, ConfigItem):
-                continue
-
-            value = item.serialize() if serialize else item.value
-            if not items.get(item.group):
-                if not item.name:
-                    items[item.group] = value
-                else:
-                    items[item.group] = {}
-
-            if item.name:
-                items[item.group][item.name] = value
-
-        return items
-
-    def save(self):
-        """ save config """
-        self._cfg.file.parent.mkdir(parents=True, exist_ok=True)
-        with open(self._cfg.file, "w", encoding="utf-8") as f:
-            json.dump(self._cfg.toDict(), f, ensure_ascii=False, indent=4)
-
-    @exceptionHandler()
-    def load(self, file=None, config=None):
-        """ load config
-
-        Parameters
-        ----------
-        file: str or Path
-            the path of json config file
-
-        config: Config
-            config object to be initialized
-        """
-        if isinstance(config, QConfig):
-            self._cfg = config
-
-        if isinstance(file, (str, Path)):
-            self._cfg.file = Path(file)
-
-        try:
-            with open(self._cfg.file, encoding="utf-8") as f:
-                cfg = json.load(f)
-        except Exception:
-            cfg = {}
-
-        # map config items'key to item
-        items = {}
-        for name in dir(self._cfg.__class__):
-            item = getattr(self._cfg.__class__, name)
-            if isinstance(item, ConfigItem):
-                items[item.key] = item
-
-        # update the value of config item
-        for k, v in cfg.items():
-            if not isinstance(v, dict) and items.get(k) is not None:
-                items[k].deserializeFrom(v)
-            elif isinstance(v, dict):
-                for key, value in v.items():
-                    key = k + "." + key
-                    if items.get(key) is not None:
-                        items[key].deserializeFrom(value)
+with contextlib.redirect_stdout(None):
+    from qfluentwidgets import QConfig, OptionsConfigItem, BoolValidator, ConfigItem, OptionsValidator
 
 
 class DSakikoConfig(QConfig):
@@ -637,12 +255,12 @@ def migrate_from_old_config(cfg: DSakikoConfig, enable_warning: bool = False):
         # 删除旧文件
         try:
             os.remove("../API_Choice.json")
-        except Exception:
+        except OSError:
             if enable_warning:
                 warnings.warn("无法删除旧的 API_Choice.json 配置文件。这可能导致新配置被旧配置反向覆盖，建议手动删除该文件。")
         try:
             os.remove("../API Key.txt")
-        except Exception:
+        except OSError:
             if enable_warning:
                 warnings.warn("无法删除旧的 API Key.txt 配置文件。这可能导致新配置被旧配置反向覆盖，建议手动删除该文件。")
 
@@ -662,7 +280,7 @@ def migrate_from_old_config(cfg: DSakikoConfig, enable_warning: bool = False):
         # 删除文件
         try:
             os.remove("../is_fp32.txt")
-        except Exception:
+        except OSError:
             if enable_warning:
                 warnings.warn("无法删除旧的 is_fp32.txt 配置文件。这可能导致新配置被旧配置反向覆盖，建议手动删除该文件。")
     except Exception:
@@ -680,7 +298,7 @@ def migrate_from_old_config(cfg: DSakikoConfig, enable_warning: bool = False):
         # 删除文件
         try:
             os.remove("../if_delete_audio_cache.txt")
-        except Exception:
+        except OSError:
             if enable_warning:
                 warnings.warn("无法删除旧的 if_delete_audio_cache.txt 配置文件。这可能导致新配置被旧配置反向覆盖，建议手动删除该文件。")
     except Exception:
@@ -698,7 +316,7 @@ def migrate_from_old_config(cfg: DSakikoConfig, enable_warning: bool = False):
         # 删除文件
         try:
             os.remove("../reference_audio/GSV_sample_rate.txt")
-        except Exception:
+        except OSError:
             if enable_warning:
                 warnings.warn("无法删除旧的 GSV_sample_rate.txt 配置文件。这可能导致新配置被旧配置反向覆盖，建议手动删除该文件。")
     except Exception:
@@ -713,7 +331,7 @@ def migrate_from_old_config(cfg: DSakikoConfig, enable_warning: bool = False):
         # 删除文件
         try:
             os.remove("../reference_audio/character_order.json")
-        except Exception:
+        except OSError:
             if enable_warning:
                 warnings.warn("无法删除旧的 character_order.json 配置文件。这可能导致新配置被旧配置反向覆盖，建议手动删除该文件。")
